strategy_report/requirements.txt
annotated-types==0.7.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachelib==0.9.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-extensions==1.0.18
dash-html-components==2.0.0
dash-table==5.0.0
dataclass-wizard==0.22.3
EditorConfig==0.12.4
Flask==3.0.3
Flask-Caching==2.3.0
frozendict==2.4.4
html5lib==1.1
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsbeautifier==1.15.1
kaleido==0.2.1
lxml==5.3.0
MarkupSafe==2.1.5
more-itertools==10.4.0
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
paramiko==3.4.1
peewee==3.17.6
platformdirs==4.3.6
plotly==5.23.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
retrying==1.3.4
ruff==0.4.10
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
sshtunnel==0.4.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.3
yfinance==0.2.43
zipp==3.19.2
